feat: updates to download resources page
