+++ apps/nextjs/src/app/aila/[id]/download/SurveyDialogLauncher.ts
@@ -1,0 +1,24 @@
+import { useEffect } from "react";
+
+import { usePosthogFeedbackSurvey } from "hooks/surveys/usePosthogFeedbackSurvey";
+
+import { useDialog } from "@/components/AppComponents/DialogContext";
+
+export function SurveyDialogLauncher() {
+  const { survey } = usePosthogFeedbackSurvey({
+    closeDialog: () => null,
+    surveyName: "End of Aila generation survey launch aug24",
+  });
+  const { setDialogWindow } = useDialog();
+
+  useEffect(() => {
+    if (survey) {
+      const timer = setTimeout(() => {
+        setDialogWindow("feedback");
+      }, 3000);
+      return () => clearTimeout(timer);
+    }
+  }, [survey, setDialogWindow]);
+
+  return null;
+}
